refactor(policy-types): Stage 0b — move cmd_vel + action-claim into kirra-policy-types
ADR-0035 Stage 0b. Completes the Stage 0 foundation begun in 0a (#919),
unblocking Stage 1 (kirra-industrial).

Moves two modules VERBATIM (behaviour-identical pure moves) out of the
root crate into the lean kirra-policy-types leaf:

- gateway::cmd_vel → kirra_policy_types::cmd_vel (CmdVel / CmdVelLimits /
  DEFAULT_CMD_VEL_LIMITS / validate_cmd_vel + its MC/DC tests). serde-only.
- the posture-aware claim cluster from src/action_filter.rs →
  kirra_policy_types::action_claim (ActionClaim / ActionDecision /
  evaluate_action_claim + its tests). This is exactly the surface the
  industrial layer (protocol_adapter) consumes, so relocating it is what
  lets kirra-industrial become a clean leaf in Stage 1.

The governor-trait-generic ActionFilter<C> / process_agent_intent STAYS in
the root crate — it depends on the SafetyContract / SafetyGovernor / etc.
traits still defined in lib.rs, whose relocation is a later ADR-0035 stage.
Confirmed protocol_adapter imports only the claim cluster, never
ActionFilter<C>, so no governor-trait move is needed here.

kirra-policy-types gains kirra-core (for the shared FleetPosture) +
serde/serde_json (the claim payload is wire-deserialized); still a lean
leaf, no service/DB deps. The root crate's gateway::cmd_vel and the claim
cluster in action_filter.rs are now pub-use re-export shims, so every
existing crate::gateway::cmd_vel::* and crate::action_filter::{ActionClaim,
ActionDecision, evaluate_action_claim} path resolves unchanged.

Verified: leaf 27 tests + root 878 lib tests green, both trees fmt-clean,
quality-guardrails satisfied, orphan-core green. All four lockfiles (root +
detached fuzz/parko/kirra-verifier-pg) updated minimally for the new dep
edges.
